int64_t pass on DSYEV_descending

+++ psi4/src/psi4/libciomr/dsyev_descending.cc
@@ -52,12 +52,12 @@ namespace psi {
                                    double* const* const e_vecs /* = nullptr*/) {
     const auto info = DSYEV_ascending(N, array, e_vals, e_vecs);
     // Reverse the order of eigenvalues
-    for (int i = 0; i < N / 2; i++) {
+    for (int64_t i = 0; i < N / 2; i++) {
         std::swap(e_vals[i], e_vals[N - i - 1]);
     }
     // Reverse the order of columns of the row-major 2D eigenvector array
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N / 2; j++) {
+    for (int64_t i = 0; i < N; i++) {
+        for (int64_t j = 0; j < N / 2; j++) {
             std::swap(e_vecs[i][j], e_vecs[i][N - j - 1]);
         }
     }
